+++ src/gevent/_socket2.py
@@ -261,7 +261,7 @@ class socket(object):
                 result = self._sock.connect_ex(address)
                 if not result or result == EISCONN:
                     break
-                elif (result in (EWOULDBLOCK, EINPROGRESS, EALREADY)) or (result == EINVAL and is_windows):
+                if (result in (EWOULDBLOCK, EINPROGRESS, EALREADY)) or (result == EINVAL and is_windows):
                     self._wait(self._write_event)
                 else:
                     raise error(result, strerror(result))