+++ src/gevent/testing/openfiles.py
@@ -108,9 +108,32 @@ else:
            (socket.listen(1)). Unlike the lsof implementation, this will only
            return sockets in a state like that.
         """
+        # We've seen OSError: No such file or directory
+        # /proc/PID/fd/NUM. This occurs in the loop that checks open
+        # files. It first does listdir() and then tries readlink() on
+        # each file. But the file went away. This must be because of
+        # async GC in PyPy running destructors at arbitrary times.
+        # This became an issue in PyPy 7.2 but could theoretically be
+        # an issue with any objects caught in a cycle. Try to clean
+        # that up before we begin.
+        import gc
+        gc.collect()
+        gc.collect()
         results = dict()
-        process = psutil.Process()
-        results['data'] = process.open_files() + process.connections('all')
+        gc.disable()
+        try:
+            for _ in range(3):
+                try:
+                    process = psutil.Process()
+                    results['data'] = process.open_files() + process.connections('all')
+                    break
+                except OSError:
+                    pass
+            else:
+                # No break executed
+                raise unittest.SkipTest("Unable to read open files")
+        finally:
+            gc.enable()
         for x in results['data']:
             results[x.fd] = x
         results['data'] += ['From psutil', process]